@netrunnereve 2025-02-22
> Perf is about neutral on average, maybe down a tiny bit?

Interesting. Let's wait for some more results.